Only include the TOP 3 items of the stack trace. Only quit the game when the offending thread is GMS (AdMob)

+++ Assets/Plugins/Android/ANRSupervisor/ANRSupervisor.java
@@ -179,7 +179,76 @@ class ANRSupervisorRunnable implements Runnable
 					{
 						ANRSupervisor.Log("Thread " + this.mHandler.getLooper() + " DID NOT respond within " + mTimeoutCheck + " seconds");
 
-						String report = getProcessJson(this.mHandler.getLooper().getThread());
+						ByteArrayOutputStream bos = new ByteArrayOutputStream();
+						PrintStream ps = new PrintStream(bos);
+
+						// Get all stack traces in the system
+						Map<Thread, StackTraceElement[]> stackTraces = Thread.getAllStackTraces();
+						Locale l = Locale.getDefault();
+		
+						String deviceName = "";
+						try
+						{
+							android.content.ContentResolver cr = com.unity3d.player.UnityPlayer.currentActivity.getApplicationContext().getContentResolver();
+							deviceName = android.provider.Settings.Secure.getString(cr, "device_name");
+							if (deviceName == null || deviceName.length() <= 0)
+							{
+								deviceName = android.provider.Settings.Secure.getString(cr, "bluetooth_name");
+							}
+						}
+						catch (Exception e) {}
+
+						ps.print(String.format(l, "{\"title\":\"ANR Report\",\"build_version\":\"%s\",\"device\":\"%s\",\"name\":\"%s\",\"callstacks\":[",
+							String.valueOf(BuildConfig.VERSION_NAME), android.os.Build.FINGERPRINT, deviceName));
+
+						Thread supervisedThread = this.mHandler.getLooper().getThread();
+						boolean isFirstThread = true;
+						boolean gmsThreadIsBlocked = false;
+						for (Thread thread : stackTraces.keySet())
+						{
+							boolean isBlocked = thread.getState().equals("BLOCKED");
+							if (thread == supervisedThread ||
+								thread.getName().equals("main") ||
+								thread.getName().equals("UnityMain") ||
+								isBlocked)
+							{
+								if (isFirstThread) { isFirstThread = false; } else { ps.print(","); }
+								ps.print(String.format(l, "{\"name\":\"%s\",\"state\":\"%s\"", thread.getName(), thread.getState()));
+				
+								if (thread == supervisedThread)
+								{
+									ps.print(",\"supervised\":true");
+								}
+
+								StackTraceElement[] stack = stackTraces.get(thread);
+								if (stack.length > 0)
+								{
+									ps.print(",\"stack\":[");
+									boolean isFirstLine = true;
+									int numStackLines = Math.min(stack.length, 3);
+									for (int i = 0; i < numStackLines; ++i)
+									{
+										if (isFirstLine) { isFirstLine = false; } else { ps.print(","); }
+										StackTraceElement element = stack[i];
+										ps.print(String.format(l, "{\"func\":\"%s.%s\",\"file\":\"%s\",\"line\":%d}",
+												element.getClassName(),
+												element.getMethodName(),
+												element.getFileName(), 
+												element.getLineNumber()));
+
+										if (isBlocked && element.getClassName().contains("gms.ads"))
+										{
+											gmsThreadIsBlocked = true;
+										}
+									}
+									ps.print("]");
+								}
+								ps.print("}");
+							}
+						}
+						ps.print("]}");
+
+						String report = new String(bos.toByteArray());
 						ANRSupervisor.Log(report);
 
 						ANRSupervisor.Log("Sending log to Firebase");
@@ -199,18 +268,19 @@ class ANRSupervisorRunnable implements Runnable
 							Thread.sleep(100);
 						}
 
-						//ANRSupervisor.Log("Waiting another 4 seconds for the report to come through...");
-						//callback.wait(4000);
-
-						ANRSupervisor.Log("Checking for false-positive");
-						if (!callback.isCalled())
+						// When we are blocked by GMS.ADS, quit the game.
+						if (gmsThreadIsBlocked)
 						{
-							ANRSupervisor.Log("Killing myself");
-							// If the supervised thread still did not respond, quit the app.
-							android.os.Process.killProcess(android.os.Process.myPid());
+							ANRSupervisor.Log("Checking for false-positive");
+							if (!callback.isCalled())
+							{
+								ANRSupervisor.Log("Killing myself");
+								// If the supervised thread still did not respond, quit the app.
+								android.os.Process.killProcess(android.os.Process.myPid());
 
-							ANRSupervisor.Log("Exiting the app");
-							System.exit(0); // SNAFU
+								ANRSupervisor.Log("Exiting the app");
+								System.exit(0); // SNAFU
+							}
 						}
 					}
 					else
@@ -233,71 +303,6 @@ class ANRSupervisorRunnable implements Runnable
 		this.mStopCompleted = true;
 
 		ANRSupervisor.Log("supervision stopped");
-	}
-
-	public String getProcessJson(Thread supervisedThread)
-	{
-		ByteArrayOutputStream bos = new ByteArrayOutputStream();
-		PrintStream ps = new PrintStream(bos);
-
-		// Get all stack traces in the system
-		Map<Thread, StackTraceElement[]> stackTraces = Thread.getAllStackTraces();
-		Locale l = Locale.getDefault();
-		
-		String deviceName = "";
-		try
-		{
-			android.content.ContentResolver cr = com.unity3d.player.UnityPlayer.currentActivity.getApplicationContext().getContentResolver();
-			deviceName = android.provider.Settings.Secure.getString(cr, "device_name");
-			if (deviceName == null || deviceName.length() <= 0)
-			{
-				deviceName = android.provider.Settings.Secure.getString(cr, "bluetooth_name");
-			}
-		}
-		catch (Exception e) {}
-
-		ps.print(String.format(l, "{\"title\":\"ANR Report\",\"build_version\":\"%s\",\"device\":\"%s\",\"name\":\"%s\",\"callstacks\":[",
-			String.valueOf(BuildConfig.VERSION_NAME), android.os.Build.FINGERPRINT, deviceName));
-
-		boolean isFirstThread = true;
-		for (Thread thread : stackTraces.keySet())
-		{
-			if (thread == supervisedThread ||
-				thread.getName().equals("main") ||
-				thread.getName().equals("UnityMain") ||
-				thread.getState().equals("BLOCKED"))
-			{
-				if (isFirstThread) { isFirstThread = false; } else { ps.print(","); }
-				ps.print(String.format(l, "{\"name\":\"%s\",\"state\":\"%s\"", thread.getName(), thread.getState()));
-				
-				if (thread == supervisedThread)
-				{
-					ps.print(",\"supervised\":true");
-				}
-
-				StackTraceElement[] stack = stackTraces.get(thread);
-				if (stack.length > 0)
-				{
-					ps.print(",\"stack\":[");
-					boolean isFirstLine = true;
-					for (int i = 0; i < stack.length; ++i)
-					{
-						if (isFirstLine) { isFirstLine = false; } else { ps.print(","); }
-						StackTraceElement element = stack[i];
-						ps.print(String.format(l, "{\"func\":\"%s.%s\",\"file\":\"%s\",\"line\":%d}",
-								element.getClassName(),
-								element.getMethodName(),
-								element.getFileName(), 
-								element.getLineNumber()));
-					}
-					ps.print("]");
-				}
-				ps.print("}");
-			}
-		}
-		ps.print("]}");
-
-		return new String(bos.toByteArray());
 	}
 
 	private synchronized void checkStopped() throws InterruptedException
